+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/providers/LintNotificationProvider.kt
@@ -1,7 +1,9 @@
 package com.github.blarc.gitlab.template.lint.plugin.providers
 
+import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle.message
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils
+import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintErrorsEnum
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pipeline
 import com.github.blarc.gitlab.template.lint.plugin.settings.ProjectSettings
 import com.intellij.codeInsight.hint.HintUtil
@@ -31,6 +33,18 @@ class LintNotificationProvider : EditorNotificationProvider {
                     errors?.joinToString(separator = " ") ?: "Lint response was invalid, but no errors were returned."
             }
 
+            panel.text = when (errors?.firstOrNull()) {
+                GitlabLintErrorsEnum.REFERENCE_NOT_FOUND.message -> message("lint.error.reference-not-found")
+                GitlabLintErrorsEnum.PIPELINE_FILTERED_OUT.message -> {
+                    panel.createActionLabel(message("lint.error.explanation")) {
+                        GitlabLintBundle.openPipelineFilteredIssue()
+                    }
+                    message("lint.error.pipeline-filtered")
+                }
+                else -> errors?.joinToString(separator = " ")
+                    ?: "Lint response was invalid, but no errors were returned."
+            }
+
             panel.createActionLabel(message("lint.error.ignore")) {
                 project.service<ProjectSettings>().ignoredErrors.getOrPut(file.path) { mutableSetOf() }
                     .add(errors?.get(0) ?: "")
@@ -41,5 +55,4 @@ class LintNotificationProvider : EditorNotificationProvider {
         }
         return@Function null
     }
-
 }